Keep the notification bell in the top bar on mobile
On narrow viewports the nav links collapse behind the hamburger, and
NotificationsNav lived inside that collapsed menu. The unread count was
fetched and stored, but the badge was invisible until the menu was
opened, so there was no way to tell you had notifications at a glance.

Render the bell in the nav bar itself on mobile, beside the hamburger,
and drop the Notifications entry from the menu. Desktop is unchanged:
the bell stays with the other nav links.

The two placements are separate DOM positions, so the component is
rendered in both and a single media query in App.vue picks which copy is
visible - the same block that toggles the hamburger, so the two cannot
drift apart. The unread count fetch moves up to App.vue accordingly, to
keep it at one request regardless of how many bells are mounted.

The bell is now icon-only in both layouts, which removes the mobile
label that the badge used to overlap, along with the duplicated 768px
breakpoint that label needed. Its accessible name is translated instead
of hardcoded English, since that name is all a screen reader now has.
