+++ packages/vue-client/src/locales/de.json
@@ -7,6 +7,7 @@
   "logout": "Abmelden",
   "admin": "Administration",
   "notifications": "Nachrichten",
+  "notifications-unread": "{count} ungelesen",
   "pagination": {
     "size-label": "Anzahl:",
     "first": "Zum Anfang",
