+++ packages/vue-client/src/locales/en.json
@@ -7,6 +7,7 @@
   "logout": "Log out",
   "admin": "Admin",
   "notifications": "Notifications",
+  "notifications-unread": "{count} unread",
   "pagination": {
     "size-label": "show:",
     "first": "First",
